fix(M2): address CodeHawk audit findings and self-review issues from PR #2
CodeHawk findings fixed (all 17 real items of 19; #17 and #2a are false positives,
#15 is a non-definding 'no issue found' confirmation):

- pipeline: stage2_cursor no longer overloaded with TLM frame sequence — new
  checkpoint column stage1_tlm_seq (migration 0004) separates the TLM stage-1
  frame sequence from the stage-2 staging cursor. Previously every TLM file
  silently lost all records in stage 2 (cursor = last frame seq skipped all
  staged rows). (CH#3 = self#1, Critical)
- tlm_parser: refill() now APPENDS after the unconsumed tail instead of reading
  from the head, which overwrote the window for any stream > 64 KB. (CH#7,
  Critical)
- tlmcodec: consumed and frame_info are set on every post-header outcome
  (BAD_FLAGS, PAYLOAD_TOO_LARGE, CRC, TLV, compression included), eliminating
  the infinite SkipRecord loop for frames with reserved flags or oversized
  payloads. (self#2, Critical — CH missed this path)
- windows: compute_stats propagates errors via Result instead of silently
  returning zero stats that were persisted. (CH#9/#10)
- windows: supersede + corrected version INSERT are now in one transaction.
  (CH#11 = self#3)
- windows: closed_ cache evicted beyond 100k entries (bounded memory).
- pipeline: seed query failure now propagates (was silently swallowed,
  disabling late detection). (CH#4)
- pipeline: EOF tail drain collects kLate and applies corrections. (CH#5 = self#5)
- missing: dead code in gap<=0 branch removed. (CH#1)
- config: duplicate (device, metric) calibration entries rejected. (CH#16)
- watermark: normal drain runs before the cap check so on-time samples are not
  mislabelled kForcedFlush. (CH#18)
- config: expected_period must be positive (prevents SIGFPE division by zero
  in compute_stats). (CH#8 / #2b)

Self-review findings fixed (not in CodeHawk):
- test_consistency: fixture files are copied to the run's input dir before
  processing (the pipeline archives consumed files and would otherwise
  destroy the committed fixtures on each test run).
- tlm_parser: truncated-trailing-frame path drops the window so the next
  call returns EOF instead of looping.
- tlmcodec: TLM_MAX_PAYLOAD and related size macros use size_t to prevent
  implicit widening in pointer arithmetic.

Verification: 132 test cases / 975 assertions green on Release and
ASan+UBSan; clang-tidy clean; clang-format applied.

+++ c/tlmcodec/tlmcodec.h
@@ -41,16 +41,16 @@ typedef enum tlm_status {
     TLM_ERR_INVALID_ARGUMENT = 14   /* NULL or otherwise invalid arguments */
 } tlm_status;
 
-#define TLM_FRAME_HEADER_SIZE 16u
-#define TLM_FILE_HEADER_SIZE 24u
-#define TLM_MAX_PAYLOAD (4u * 1024u * 1024u)
-#define TLM_MAX_TAGS 16u
-#define TLM_TAG_KEY_MAX 128u
-#define TLM_TAG_VALUE_MAX 128u
-#define TLM_ID_MAX 64u
-#define TLM_UNIT_MAX 32u
-#define TLM_MAX_TLV_COUNT 256u
-#define TLM_MAX_TLV_VALUE 4096u
+#define TLM_FRAME_HEADER_SIZE ((size_t)16)
+#define TLM_FILE_HEADER_SIZE ((size_t)24)
+#define TLM_MAX_PAYLOAD ((size_t)4 * 1024 * 1024)
+#define TLM_MAX_TAGS ((size_t)16)
+#define TLM_TAG_KEY_MAX ((size_t)128)
+#define TLM_TAG_VALUE_MAX ((size_t)128)
+#define TLM_ID_MAX ((size_t)64)
+#define TLM_UNIT_MAX ((size_t)32)
+#define TLM_MAX_TLV_COUNT ((size_t)256)
+#define TLM_MAX_TLV_VALUE ((size_t)4096)
 
 /* Frame types. */
 #define TLM_FRAME_DATA 1u
@@ -112,11 +112,16 @@ typedef struct tlm_record {
 tlm_status tlm_decode_file_header(const uint8_t* buf, size_t len, tlm_file_header* out);
 
 /* Parses one frame starting at buf[0].
- * On TLM_OK: *consumed holds the total frame length; *info is filled; for data frames
+ * On success: *consumed holds the total frame length; *info is filled; for data frames
  * *record is filled. Unknown frame types (not 1/2/3) parse their framing but leave the
  * record untouched. Sequence state is advanced only on accepted frames; a regression
  * returns TLM_ERR_SEQUENCE_REGRESSION with *consumed set so the caller can skip the
- * offending frame and continue with the next one. */
+ * offending frame and continue with the next one.
+ * Contract: once the 16-byte frame header is decoded and the full frame is present in
+ * the buffer, *consumed and *info are set on EVERY outcome — including bad flags,
+ * payload-too-large, CRC mismatch, sequence regression and TLV errors — so a streaming
+ * caller can always skip the frame and make progress. Only BAD_SYNC, TRUNCATED and
+ * INVALID_ARGUMENT leave *consumed untouched (there is no complete frame to skip). */
 tlm_status tlm_parse_frame(const uint8_t* buf, size_t len, tlm_parser_state* st,
                            tlm_frame_info* info, size_t* consumed, tlm_record* record);
